+++ packages/sdk/src/proof/types.ts
@@ -5,7 +5,7 @@ export interface BaseProof {
   pi_b: any;
   pi_c: any;
   protocol: string;
-  curve?: any;
+  curve?: any; // TODO: make this not optional
 }
 
 export interface MerkleProofInput {