+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+- add function `JoinSplitPublicSignalsToArray`
+- add function `unpackFromSolidityProof` to convert `SolidityProof -> BaseProof`
 - factor contents of `NocturneContext.proveJoinSplitTx` into a standalone helper
 - Add `makeProvenJoinSplitTx` util function that removes additional fields `PreProofJoinSplitTx` has
 - Fix `NocturneContext.gatherMinimumNotes` bug by adding missing await for `this.ensureMinimumForAssetRequest` and updating test